BJData optimized binary array type
Introduces a dedicated `B` marker for bytes. This is used as the strong
type marker in optimized array format to encode binary data such that
it can also be decoded back to binary data (instead of decoding as an
integer array).

See https://github.com/NeuroJSON/bjdata/pull/6 for further information.